♻️ refactor(profile): extract ProfileStats, AvatarUpload, and DisplayNameForm components
